chore: add scanner for retrieving tokens
Adding a scanner. This scanner be provisioned with a stream upon which the scanner will use a generator to yield tokens found in the stream in sequence of placement in the string.

By default the scanner will remove whitespace tokens from the stream, but these can be incorporated in the response by enabling the option.